Begginings of node discovery and task engine.
Updates to redfish library to create from a serialized form of it.

Adds endpoint to begin ingestion process from range of IPs.

Update base DB structure to include ingest task uuid column.

The task engine uses bullmq to handle the processing of a workqueue and custom
built task scheduler and workflow dependency tracker. The workflows are stored
in memory at this time. Eventually need to persist that to handle app restarts.